+++ Source/DxManagerRender.cpp
@@ -656,7 +656,7 @@ void DxManager::RenderCascadedShadowMap()
 HRESULT DxManager::Render()
 {
 	if(this->RendererSleep > 0)
-		Sleep(this->RendererSleep);
+		Sleep((DWORD)this->RendererSleep);
 
 	// Timer
 	LARGE_INTEGER li;
@@ -718,7 +718,7 @@ HRESULT DxManager::Render()
 	//for(int q = 0; q < this->lights.size(); q++)
 		//DrawScreenSpaceBillboardDebug(this->Dx_DeviceContext, this->Shader_BillBoard, this->lights[q]->GetShadowMapSRV(), q); 
 	//for(int q = 0; q < this->csm->GetNrOfCascadeLevels(); q++)
-	//	DrawScreenSpaceBillboardDebug(this->Dx_DeviceContext, this->Shader_BillBoard, this->csm->GetShadowMapSRV(q), q); 
+		//DrawScreenSpaceBillboardDebug(this->Dx_DeviceContext, this->Shader_BillBoard, this->csm->GetShadowMapSRV(q), q);
